+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -11,6 +11,7 @@ export * from './Button';
 export * from './Icon';
 export * from './IconButton';
 export * from './DropdownMenu';
+export * from './ExposedDropdownMenuPicker';
 export * from './Divider';
 export * from './Host';
 export * from './LazyColumn';